pyproject.toml
[tool.poetry]
name = "ComicRestful-backend"
version = "0.0.1"
description = "Flask restful backend for Comic Generator"
authors = ["lzq910123@gmail.com"]


[tool.poetry.dependencies]
python = ">=3.9,<3.12"
flask = "^2.2.2"
flask_sqlalchemy = "^2.5.1"
flask_jwt_extended = "^4.4.4"
alembic = "^1.4.2"
gunicorn = "^20.1.0"
tomlkit = "^0.6.0"
flask_cors = "^3.0.10"
webargs = "^8.2.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.1"
scipy = "^1.9.1"
scikit-learn = "^1.1.2"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

